+++ sql2json/__main__.py
@@ -8,11 +8,20 @@ from .sql2json import run_query2json
 
 
 def handle_run_query2json(
-    name="default", query="default", wrapper=False, first=False, key=False, **kwargs
+    name="default",
+    query="default",
+    wrapper=False,
+    first=False,
+    key="",
+    value="",
+    jsonkeys="",
+    **kwargs
 ):
-    result = run_query2json(name, query, wrapper, first, key, **kwargs)
+    result = run_query2json(name, query, wrapper, first, key, value, jsonkeys, **kwargs)
 
-    if key and first:
+    if key and value and first:
+        print(json.dumps(result))
+    elif key and first:
         print(result)
     else:
         print(json.dumps(result))